VRAM Calculator! (first finished app!!)
I am extremely excited to be posting this as my first ever c# app! I don't really need to say anything else, it calculates the amount of VRAM you need for a resolution at a certain refresh rate, then gives you a recommended amount of VRAM (literally just multiplying the VRAM calculated before time 2 lol).

I came up with this while I was calculating the perfect render resolution setting in SteamVR (since I have a GTX 1060 3GB this was very useful to me actually), but it was taking an extremely long time to figure out the resolution then put it through allll of the calculations on the windows calculator and stuff, so I thought "why don't I just make a program that does this for me?". so I did, and this is the final product.